+++ compiler/ml/unified_ops.ml
@@ -149,6 +149,39 @@ let entries =
         };
     };
     {
+      path = builtin "<<";
+      name = "%lsl";
+      form = Binary;
+      specialization =
+        {
+          int = Plslint;
+          bool = None;
+          float = None;
+          bigint = Some Plslbigint;
+          string = None;
+        };
+    };
+    {
+      path = builtin ">>";
+      name = "%asr";
+      form = Binary;
+      specialization =
+        {
+          int = Pasrint;
+          bool = None;
+          float = None;
+          bigint = Some Pasrbigint;
+          string = None;
+        };
+    };
+    {
+      path = builtin ">>>";
+      name = "%lsr";
+      form = Binary;
+      specialization =
+        {int = Plsrint; bool = None; float = None; bigint = None; string = None};
+    };
+    {
       path = builtin "mod";
       name = "%mod";
       form = Binary;