+++ tests/tests/src/belt_int_test.mjs
@@ -37,6 +37,9 @@ Mocha.describe("Belt_int_test", () => {
     Test_utils.eq("File \"belt_int_test.res\", line 42, characters 7-14", 0, 0);
     Test_utils.eq("File \"belt_int_test.res\", line 43, characters 7-14", 0, 0);
     Test_utils.eq("File \"belt_int_test.res\", line 44, characters 7-14", 1, 1);
+    Test_utils.eq("File \"belt_int_test.res\", line 45, characters 7-14", 16, 16);
+    Test_utils.eq("File \"belt_int_test.res\", line 46, characters 7-14", 2, 2);
+    Test_utils.eq("File \"belt_int_test.res\", line 47, characters 7-14", 2, 2);
   });
 });